requirements.txt
asgiref
certifi
charset-normalizer
Django
django-environ
django-filter
djangorestframework
idna
Markdown
mysql-connector-python
requests
setuptools
sqlparse
stripe
typing_extensions
urllib3
wheel
django-redis
pillow
paramiko
sshtunnel
django-cors-headers
google-auth
firebase-admin
